requirements.txt
pysensors
pysqlite3
